**libc, user, kernel: improve type safety, nullptr usage, Makefiles, and add clang-tidy enhancements**
- Replaced `NULL` with `nullptr` across the codebase for improved type safety and modern C++ style compliance.
- Enhanced type handling by properly casting types such as `ssize_t` where necessary to prevent potential truncation or overflow issues.
- Refined Makefiles: removed redundant custom libc object builds for `doom`; now uses prebuilt `libc.a` from the main build for simplicity.
- Added `SKIP_PATTERNS` in `clang_tidy_all.py` to exclude third-party and legacy code paths from clang-tidy checks.
- Improved debug flags, file skipping, and code style, addressing compiler and static analysis warnings using appropriate annotations (`NOLINT`, etc.).

+++ user/cat.c
@@ -4,10 +4,10 @@ char buf[512];
 
 void cat(int fd)
 {
-    int n;
+    ssize_t n;
 
     while ((n = read(fd, buf, sizeof(buf))) > 0) {
-        if (write(1, buf, n) != n) {
+        if (write(1, buf, (size_t)n) != n) {
             printf("cat: write error\n");
             exit();
         }
